Split functionality in multiple modules
